rfctr(email): final renaming of new_email modules
